pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.21.0",
    "apscheduler>=3.11.0",
    "python-dotenv>=1.1.1",
]
